now box sticks to head, new script file
optimized file structure, make robot have a movment script and a attachment script so it wont be a huge long single script.
