+++ My project/Assets/Scripts/AttachmentHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 10374c912c71010439b2dc54cd54d59e
